feat!: add option to change the logo
added a toggle to change the intire logo
add a field to link an big and a small logo

BREAKING CHANGE: We have revised the markup of the header area. The structures were divided into other partials and these were renamed so that they are closer to what they actually do. In addition, the playback of the two new images was added, see Resources/Private/Partials/Page/Header.html and Resources/Private/Partials/Page/Header.Image-Trademark.html.

Refs: ITZBUNDPHP-2903