feat: Implement feature flags and color configuration
Implemented feature flags for improved control over active features within the project and introduced a new color configuration feature. Updated the 'sites.php' and 'README.md' files to include details regarding the usage of feature flags in PHP and Fluid templates, as well as instructions for color configurations.

Additionally, the 'FeatureFlagViewHelper' class was created to evaluate the state of a feature flag. The relevant ViewHelper namespace was also included in the 'Colors.html' Fluid template to conditionally render based on the 'ITZBUNDPHP-2877' feature flag.

Refs: ITZBUNDPHP-2877